backend/requirements.txt
# FastAPI
fastapi
uvicorn[standard]
python-multipart

# LangChain Core
langchain
langchain-core
langchain-community
langchain_openai


#rag-basics
sentence-transformers
scikit-learn
pymilvus

# Embeddings & LLMs
openai

# Utilities
pydantic
pydantic-settings
python-dotenv
pypdf

# Optional: Other useful packages
langsmith